Make OMR::MethodBuilder client aware
Add MethodBuilder RequestFunction client callback support

Defining the client API for JitBuilder includes providing a mechanism for
all the call backs used by JitBuilder to be used by clients via the
client API. This mechanism is a C callback for all virtual functions
that can be overidden. For MethodBuilder, that is RequestFunction()
which returns a bool. The callback can be set via
setClientCallback_RequestFunction() and will be called with the client
object and the requested function name as parameters.

Remove OMRMethodBuilder::injectIL() as superfluous

All it does is to call the superclass's version and return its
return value, so we should just remove it.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>