Add client pointer to IlBuilder
Move OrphanBytecodeBuilder from MethodBuilder to IlBuilder

Only being able to create BytecodeBuilders from inside a MethodBuilder
is needlessly limiting; IlBuilder is the super class, so let's move it
there for maximum flexibility.

Add IlBuilder buildIL client callback support

Defining the client API for JitBuilder includes providing a mechanism for
all the call backs used by JitBuilder to be used by clients via the
client API. This mechanism is a C callback for all virtual functions
that can be overidden. For IlBuilder, that is buildIL() which returns
a bool. The callback can be set via setClientCallback_buildIL() and
will be called with the client object as parameter.

Properly initialize _client* fields in OMRIlBuilder

These fields need to be initialized to 0 to properly detect when a
new client object is needed / callback should be invoked.

Don't use varargs in OMR::IlBuilder::ifAnd() and ifOr()

After separating the client API from the implementation, these two
implementation services should no longer use varargs: the client
API takes care of translating the varargs parameters into arrays
in order to translate client objects to implementation objects.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>

fixes for OMRIlBuilder

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>