Add initial C JitBuilder API generator
The C generator is a class that extends the C++ generator, overriding
certain helpers to emit C code instead of C++. In its current state, the
generator produces C header files and corresponding C++ source files.
However, only the header files are correct because the source file
generation functionality is not yet overriden.

Signed-off-by: Leonardo Banderali <leonardo2718@protonmail.com>