Properly reinitialize register candidate region in GRA
When GRA releases the current register candidates, it does so by
explicitly calling the destructor of the Region used to allocate those
candidates. However, when reinitializing the Region, the could would
call the `operator new` overload that takes a Region by reference,
resulting in a new region being allocated inside the destructed region.
This commit corrects the code to instead do a placement new on top of
the destructed Region.

Signed-off-by: Leonardo Banderali <leonardo2718@protonmail.com>