Implement lowering for newvalue
Right now, `newvalue` always gets lowered to a `new`. The first child of
the `newvalue` becomes the child of the `new`. The other children are
moved before the `new` and individually anchored. The allocated object
is initialized by indirect stores immediately following the `new` of the
anchored values.

Put newvalue lowering in its own function

Move lowering point to just after the opcodes that must be inlined

The previous point where newvalue was being lowered didn't make much
sense because it was with other things that are not related to lowering.
The new situation is not ideal because it still stands out, but at least
the parts that actually do lowering are together.

Signed-off-by: Leonardo Banderali <leonardo2718@protonmail.com>